rename dbhelper and remove unused code
